(Chore) rename dailyChart to liveChart
